Move GraphClient and related classes under Graph namespace.
