+++ Buy/Graph.QueryError.swift
@@ -1,5 +1,5 @@
 //
-//  GraphError.swift
+//  Graph.QueryError.swift
 //  Buy
 //
 //  Created by Shopify.
@@ -26,26 +26,28 @@
 
 import Foundation
 
-public enum GraphError: Error {
-    
-    public struct Reason {
+public extension Graph {
+    public enum QueryError: Error {
         
-        let message: String
-        let line:    Int?
-        let column:  Int?
-        
-        init(json: JSON) {
-            self.message = (json["message"] as? String) ?? "Unknown error"
-            self.line    = json["line"]     as? Int
-            self.column  = json["column"]   as? Int
+        public struct Reason {
+            
+            let message: String
+            let line:    Int?
+            let column:  Int?
+            
+            init(json: JSON) {
+                self.message = (json["message"] as? String) ?? "Unknown error"
+                self.line    = json["line"]     as? Int
+                self.column  = json["column"]   as? Int
+            }
         }
+        
+        case request(error: Error?)
+        case http(statusCode: Int)
+        case noData
+        case jsonDeserializationFailed(data: Data?)
+        case invalidJson(json: Any)
+        case schemaViolationError(violation: SchemaViolationError)
+        case queryError(reasons: [Reason])
     }
-    
-    case request(error: Error?)
-    case http(statusCode: Int)
-    case noData
-    case jsonDeserializationFailed(data: Data?)
-    case invalidJson(json: Any)
-    case schemaViolationError(violation: SchemaViolationError)
-    case queryError(reasons: [Reason])
 }
